ProfileOverview refactoring
